@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.SignalR.Client;
@@ -8,12 +7,12 @@ namespace SignalR.EventAggregatorProxy.Client.DotNetCore.Bootstrap.Factories
 {
     public interface IHubProxyFactory
     {
-        Task<IHub> Create(string hubUrl, Action<HubConnection> configureConnection, Func<IHub, Task> onStarted, Func<Task> reconnected, Action<Exception> faulted, Action connected);
+        Task<IHub> Create(string hubUrl, Func<IHub, Task> onStarted, Func<Task> reconnected, Action<Exception> faulted, Action connected, Action<HubConnection>? configureConnection);
     }
 
     public interface IHub
     {
         IDisposable On<T>(string methodName, Action<T> handler);
-        Task InvokeAsync(string methodName, object[] args, CancellationToken cancellationToken = default(CancellationToken));
+        Task InvokeAsync(string methodName, object[] args, CancellationToken cancellationToken = default);
     }
 }
