@@ -1,21 +1,21 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 
 namespace SignalR.EventAggregatorProxy.Extensions
 {
     public static class CollectionExtensions
     {
-        public static IEnumerable<T> ForEach<T>(this IEnumerable<T> collection, Action<T> action)
+        public static void ForEach<T>(this IEnumerable<T> collection, Action<T> action)
         {
-            if (collection != null)
-                foreach (var item in collection)
-                {
-                    action(item);
-                }
+            foreach (var item in collection)
+                action(item);
+        }
 
-            return collection;
+        public static IReadOnlyCollection<T> AsReadOnlyCollection<T>(this IEnumerable<T> source)
+        {
+            if (source is IReadOnlyCollection<T> collection) return collection;
+            return source.ToList();
         }
     }
 }
