@@ -8,23 +8,15 @@ using SignalR.EventAggregatorProxy.Model;
 
 namespace SignalR.EventAggregatorProxy.Hubs
 {
-    public class EventAggregatorProxyHub : Hub
+    public class EventAggregatorProxyHub(EventProxy eventProxy) : Hub
     {
-        private readonly EventProxy eventProxy;
-
-        public EventAggregatorProxyHub(EventProxy eventProxy)
-        {
-            this.eventProxy = eventProxy;
-        }
-
-
         public void Subscribe(IEnumerable<SubscriptionDto> subscriptions, bool reconnected)
         {
             if (reconnected)
                 eventProxy.UnsubscribeConnection(Context.ConnectionId);
 
             subscriptions
-                .ForEach(s => eventProxy.Subscribe(Context, s.Type, s.GenericArguments ?? new string[0], s.Constraint, s.ConstraintId));
+                .ForEach(s => eventProxy.Subscribe(Context, s.Type, s.GenericArguments ?? Array.Empty<string>(), s.Constraint, s.ConstraintId));
         }
 
         public void Unsubscribe(IEnumerable<EventType> types)
@@ -32,7 +24,7 @@ namespace SignalR.EventAggregatorProxy.Hubs
             eventProxy.Unsubscribe(Context.ConnectionId, types);
         }
 
-        public override Task OnDisconnectedAsync(Exception exception)
+        public override Task OnDisconnectedAsync(Exception? exception)
         {
             eventProxy.UnsubscribeConnection(Context.ConnectionId);
             return base.OnDisconnectedAsync(exception);
